Refactor maker home view into smaller components
